+++ lib/host/webview2gtk-host-api.h
@@ -15,6 +15,7 @@ typedef struct WebView2Host WebView2Host;
 
 WebView2Host *vala_webview2_host_create_with_xywh (void *parent_hwnd, int x, int y, int width, int height, uint16_t *url);
 void vala_webview2_host_set_bounds_xywh (WebView2Host *host, int x, int y, int width, int height);
+void vala_webview2_host_note_visible_hwnd (WebView2Host *host);
 bool vala_webview2_host_navigate (WebView2Host *host, const char *url_utf8);
 bool vala_webview2_host_navigate_to_string (WebView2Host *host, const char *html_utf8);
 void vala_webview2_host_destroy (WebView2Host *host);